fix: displayPercentage not working for inline-triggers
